Moving two update methods into after_action.go. Using UpdateTabletFields instead of UpdateTablet for mysql port change. Renaming readTablet into updateTabletFromTopo.
