+++ go/vt/tabletmanager/agent.go
@@ -34,17 +34,14 @@ import (
 	"golang.org/x/net/context"
 
 	log "github.com/golang/glog"
-	"github.com/youtube/vitess/go/event"
 	"github.com/youtube/vitess/go/history"
 	"github.com/youtube/vitess/go/netutil"
 	"github.com/youtube/vitess/go/stats"
-	"github.com/youtube/vitess/go/trace"
 	"github.com/youtube/vitess/go/vt/binlog"
 	"github.com/youtube/vitess/go/vt/dbconfigs"
 	"github.com/youtube/vitess/go/vt/health"
 	"github.com/youtube/vitess/go/vt/key"
 	"github.com/youtube/vitess/go/vt/mysqlctl"
-	"github.com/youtube/vitess/go/vt/tabletmanager/events"
 	"github.com/youtube/vitess/go/vt/tabletserver"
 	"github.com/youtube/vitess/go/vt/tabletserver/planbuilder"
 	"github.com/youtube/vitess/go/vt/tabletserver/tabletservermock"
@@ -305,24 +302,9 @@ func (agent *ActionAgent) registerQueryRuleSources() {
 	agent.QueryServiceControl.RegisterQueryRuleSource(blacklistQueryRules)
 }
 
-func (agent *ActionAgent) updateState(ctx context.Context, oldTablet *pb.Tablet, reason string) error {
-	agent.mutex.Lock()
-	newTablet := agent._tablet.Tablet
-	agent.mutex.Unlock()
-	log.Infof("Running tablet callback because: %v", reason)
-	if err := agent.changeCallback(ctx, oldTablet, newTablet); err != nil {
-		return err
-	}
-
-	event.Dispatch(&events.StateChange{
-		OldTablet: *oldTablet,
-		NewTablet: *newTablet,
-		Reason:    reason,
-	})
-	return nil
-}
-
-func (agent *ActionAgent) readTablet(ctx context.Context) (*topo.TabletInfo, error) {
+// updateTabletFromTopo will read the tablet record from the topology,
+// save it in the agent's tablet record, and return it.
+func (agent *ActionAgent) updateTabletFromTopo(ctx context.Context) (*topo.TabletInfo, error) {
 	tablet, err := agent.TopoServer.GetTablet(ctx, agent.TabletAlias)
 	if err != nil {
 		return nil, err
@@ -395,41 +377,6 @@ func (agent *ActionAgent) setTabletControl(tc *pb.Shard_TabletControl) {
 	agent.mutex.Unlock()
 }
 
-// refreshTablet needs to be run after an action may have changed the current
-// state of the tablet.
-func (agent *ActionAgent) refreshTablet(ctx context.Context, reason string) error {
-	log.Infof("Executing post-action state refresh")
-
-	span := trace.NewSpanFromContext(ctx)
-	span.StartLocal("ActionAgent.refreshTablet")
-	span.Annotate("reason", reason)
-	defer span.Finish()
-	ctx = trace.NewContext(ctx, span)
-
-	// Save the old tablet so callbacks can have a better idea of
-	// the precise nature of the transition.
-	oldTablet := agent.Tablet().Tablet
-
-	// Actions should have side effects on the tablet, so reload the data.
-	ti, err := agent.readTablet(ctx)
-	if err != nil {
-		log.Warningf("Failed rereading tablet after %v - services may be inconsistent: %v", reason, err)
-		return fmt.Errorf("Failed rereading tablet after %v: %v", reason, err)
-	}
-
-	if updatedTablet := agent.checkTabletMysqlPort(ctx, ti); updatedTablet != nil {
-		agent.mutex.Lock()
-		agent._tablet = updatedTablet
-		agent.mutex.Unlock()
-	}
-
-	if err := agent.updateState(ctx, oldTablet, reason); err != nil {
-		return err
-	}
-	log.Infof("Done with post-action state refresh")
-	return nil
-}
-
 func (agent *ActionAgent) verifyTopology(ctx context.Context) error {
 	tablet := agent.Tablet()
 	if tablet == nil {
@@ -459,7 +406,7 @@ func (agent *ActionAgent) verifyServingAddrs(ctx context.Context) error {
 // If initUpdateStream is set, update stream service will also be registered.
 func (agent *ActionAgent) Start(ctx context.Context, mysqlPort, vtPort, gRPCPort int32, initUpdateStream bool) error {
 	var err error
-	if _, err = agent.readTablet(ctx); err != nil {
+	if _, err = agent.updateTabletFromTopo(ctx); err != nil {
 		return err
 	}
 
@@ -507,7 +454,7 @@ func (agent *ActionAgent) Start(ctx context.Context, mysqlPort, vtPort, gRPCPort
 	}
 
 	// Reread to get the changes we just made
-	tablet, err := agent.readTablet(ctx)
+	tablet, err := agent.updateTabletFromTopo(ctx)
 	if err != nil {
 		return err
 	}
@@ -589,7 +536,8 @@ func (agent *ActionAgent) hookExtraEnv() map[string]string {
 }
 
 // checkTabletMysqlPort will check the mysql port for the tablet is good,
-// and if not will try to update it.
+// and if not will try to update it. It returns the modified Tablet record,
+// if it was changed.
 func (agent *ActionAgent) checkTabletMysqlPort(ctx context.Context, tablet *topo.TabletInfo) *topo.TabletInfo {
 	mport, err := agent.MysqlDaemon.GetMysqlPort()
 	if err != nil {
@@ -602,12 +550,16 @@ func (agent *ActionAgent) checkTabletMysqlPort(ctx context.Context, tablet *topo
 	}
 
 	log.Warningf("MySQL port has changed from %v to %v, updating it in tablet record", tablet.PortMap["mysql"], mport)
-	tablet.PortMap["mysql"] = mport
-	if err := agent.TopoServer.UpdateTablet(ctx, tablet); err != nil {
+	if err := agent.TopoServer.UpdateTabletFields(ctx, tablet.Alias, func(t *pb.Tablet) error {
+		t.PortMap["mysql"] = mport
+		return nil
+	}); err != nil {
 		log.Warningf("Failed to update tablet record, may use old mysql port")
 		return nil
 	}
 
+	// update worked, return the new record
+	tablet.PortMap["mysql"] = mport
 	return tablet
 }
